requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
flask-migrate>=4.0
flask-login>=0.6
flask-wtf>=1.2
wtforms[email]>=3.1
argon2-cffi>=23.1
flask-admin>=1.6
mysql-connector-python>=9.0
python-dotenv>=1.0
requests>=2.31
gunicorn>=21.2
